Rearrange files in josev2 pkg
